includes updates to freezeout finder. still need to add functionality to coarse grain surface in proper time

+++ rhic/src/InitialConditionParameters.cpp
@@ -125,12 +125,12 @@ void readInitialConditionParameters(const char* configDirectory, void * params) 
 		delimiterPos = line.find("=");
 		line = line.substr(delimiterPos + 1);
 		rapidityMean = atof(line.c_str());
-        
-        getline(cFile, line);
-        line.erase(remove_if(line.begin(), line.end(), isspace), line.end());
-        delimiterPos = line.find("=");
-        line = line.substr(delimiterPos + 1);
-        sourceType = atof(line.c_str());
+
+    getline(cFile, line);
+  	line.erase(remove_if(line.begin(), line.end(), isspace), line.end());
+    delimiterPos = line.find("=");
+    line = line.substr(delimiterPos + 1);
+    sourceType = atoi(line.c_str());
 	}
 	else std::cerr << "No configuration file  %s found for initial condition parameters\n";
 
@@ -144,6 +144,6 @@ void readInitialConditionParameters(const char* configDirectory, void * params) 
 	initCond->fractionOfBinaryCollisions = fractionOfBinaryCollisions;
 	initCond->rapidityVariance = rapidityVariance;
 	initCond->rapidityMean = rapidityMean;
-    initCond->sourceType = sourceType;
+  initCond->sourceType = sourceType;
 
 }
